+++ webgraph/src/graphs/permuted_graph.rs
@@ -7,19 +7,21 @@
 
 use crate::prelude::*;
 use lender::*;
-use sux::traits::BitFieldSlice;
+use value_traits::slices::SliceByValue;
 
 #[derive(Debug, Clone)]
 /// A wrapper applying a permutation to the iterators of an underlying graph.
 ///
 /// Note that nodes are simply remapped: thus, neither the iterator on the graph
 /// nor the successors are sorted.
-pub struct PermutedGraph<'a, G: SequentialGraph, P: BitFieldSlice<usize> + ?Sized> {
+pub struct PermutedGraph<'a, G: SequentialGraph, P: SliceByValue<Value = usize> + ?Sized> {
     pub graph: &'a G,
     pub perm: &'a P,
 }
 
-impl<G: SequentialGraph, P: BitFieldSlice<usize>> SequentialLabeling for PermutedGraph<'_, G, P> {
+impl<G: SequentialGraph, P: SliceByValue<Value = usize>> SequentialLabeling
+    for PermutedGraph<'_, G, P>
+{
     type Label = usize;
     type Lender<'b>
         = Iter<'b, G::Lender<'b>, P>
@@ -45,8 +47,11 @@ impl<G: SequentialGraph, P: BitFieldSlice<usize>> SequentialLabeling for Permute
     }
 }
 
-impl<'b, G: SequentialGraph + SplitLabeling, P: BitFieldSlice<usize> + Send + Sync + Clone>
-    SplitLabeling for PermutedGraph<'b, G, P>
+impl<
+        'b,
+        G: SequentialGraph + SplitLabeling,
+        P: SliceByValue<Value = usize> + Send + Sync + Clone,
+    > SplitLabeling for PermutedGraph<'b, G, P>
 where
     for<'a> <G as SequentialLabeling>::Lender<'a>: Clone + ExactSizeLender + Send + Sync,
 {
@@ -64,9 +69,12 @@ where
     }
 }
 
-impl<G: SequentialGraph, P: BitFieldSlice<usize>> SequentialGraph for PermutedGraph<'_, G, P> {}
+impl<G: SequentialGraph, P: SliceByValue<Value = usize>> SequentialGraph
+    for PermutedGraph<'_, G, P>
+{
+}
 
-impl<'a, 'b, G: SequentialGraph, P: BitFieldSlice<usize>> IntoLender
+impl<'a, 'b, G: SequentialGraph, P: SliceByValue<Value = usize>> IntoLender
     for &'b PermutedGraph<'a, G, P>
 {
     type Lender = <PermutedGraph<'a, G, P> as SequentialLabeling>::Lender<'b>;
@@ -87,7 +95,7 @@ pub struct Iter<'node, I, P> {
 impl<
         'succ,
         I: Lender + for<'next> NodeLabelsLender<'next, Label = usize>,
-        P: BitFieldSlice<usize>,
+        P: SliceByValue<Value = usize>,
     > NodeLabelsLender<'succ> for Iter<'_, I, P>
 {
     type Label = usize;
@@ -97,21 +105,23 @@ impl<
 impl<
         'succ,
         I: Lender + for<'next> NodeLabelsLender<'next, Label = usize>,
-        P: BitFieldSlice<usize>,
+        P: SliceByValue<Value = usize>,
     > Lending<'succ> for Iter<'_, I, P>
 {
     type Lend = (usize, <Self as NodeLabelsLender<'succ>>::IntoIterator);
 }
 
-impl<L: Lender + for<'next> NodeLabelsLender<'next, Label = usize>, P: BitFieldSlice<usize>> Lender
-    for Iter<'_, L, P>
+impl<
+        L: Lender + for<'next> NodeLabelsLender<'next, Label = usize>,
+        P: SliceByValue<Value = usize>,
+    > Lender for Iter<'_, L, P>
 {
     #[inline(always)]
     fn next(&mut self) -> Option<Lend<'_, Self>> {
         self.iter.next().map(|x| {
             let (node, succ) = x.into_pair();
             (
-                self.perm.get(node),
+                self.perm.index_value(node),
                 Succ {
                     iter: succ.into_iter(),
                     perm: self.perm,
@@ -123,7 +133,7 @@ impl<L: Lender + for<'next> NodeLabelsLender<'next, Label = usize>, P: BitFieldS
 
 impl<
         L: ExactSizeLender + for<'next> NodeLabelsLender<'next, Label = usize>,
-        P: BitFieldSlice<usize>,
+        P: SliceByValue<Value = usize>,
     > ExactSizeLender for Iter<'_, L, P>
 {
     fn len(&self) -> usize {
@@ -137,15 +147,15 @@ pub struct Succ<'a, I: Iterator<Item = usize>, P> {
     perm: &'a P,
 }
 
-impl<I: Iterator<Item = usize>, P: BitFieldSlice<usize>> Iterator for Succ<'_, I, P> {
+impl<I: Iterator<Item = usize>, P: SliceByValue<Value = usize>> Iterator for Succ<'_, I, P> {
     type Item = usize;
     #[inline(always)]
     fn next(&mut self) -> Option<Self::Item> {
-        self.iter.next().map(|succ| self.perm.get(succ))
+        self.iter.next().map(|succ| self.perm.index_value(succ))
     }
 }
 
-impl<I: ExactSizeIterator<Item = usize>, P: BitFieldSlice<usize>> ExactSizeIterator
+impl<I: ExactSizeIterator<Item = usize>, P: SliceByValue<Value = usize>> ExactSizeIterator
     for Succ<'_, I, P>
 {
     #[inline(always)]